+++ packages/node-core/CHANGELOG.md
@@ -5,6 +5,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Update `ParentProject` to use `untilBlock` as and alias for `block` (#2235)
+
+## [7.2.0] - 2024-01-30
+### Changed
+- Update `@subql/apollo-links` and add specific logger for it
+
+## [7.1.0] - 2024-01-25
+### Added
+- Support for CSV output with flag `--csv-out-dir` (#2224)
+- Support for Schema Migration with `--allow-schema-migration` flag (#2179)
+
+### Changed
+- Improve modulo filter performance when there are other data sources (#2152)
+
+### Fixed
+- missing schema migration models (#2226)
 
 ## [7.0.8] - 2024-01-10
 ### Fixed
@@ -538,7 +555,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Move blockchain agnostic code from `node` to `node-core` package. (#1222)
 
-[Unreleased]: https://github.com/subquery/subql/compare/node-core/7.0.8...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/node-core/7.2.0...HEAD
+[7.2.0]: https://github.com/subquery/subql/compare/node-core/7.1.0...node-core/7.2.0
+[7.1.0]: https://github.com/subquery/subql/compare/node-core/7.0.8...node-core/7.1.0
 [7.0.8]: https://github.com/subquery/subql/compare/node-core/7.0.7...node-core/7.0.8
 [7.0.7]: https://github.com/subquery/subql/compare/node-core/7.0.6...node-core/7.0.7
 [7.0.6]: https://github.com/subquery/subql/compare/node-core/7.0.5...node-core/7.0.6